feat: add Pydantic validation, pyright/mypy, and sandbox lifecycle examples
Validate control-plane and dataplane JSON with Pydantic, add py.typed and
strict pyright/mypy CI checks, drop basic async/sync examples in favor of
sandbox lifecycle and controller CLI flows, and move governance docs to .github/.

+++ src/neevai/transport/retry.py
@@ -13,7 +13,7 @@ def calculate_backoff(attempt: int) -> float:
     """
     base = min(0.250 * (2**attempt), 8.0)
     # Match TS jitter: base * (0.5 + random * 0.5)
-    return base * (0.5 + random.random() * 0.5)
+    return float(base * (0.5 + random.random() * 0.5))
 
 
 def parse_retry_after(header: str | None) -> float | None:
